% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.rotation.R
\name{gsea.rotation}
\alias{gsea.rotation}
\title{GSEA rotation}
\usage{
gsea.rotation(X, y = NULL, S, nrot = 500, fun = t.90th)
}
\arguments{
\item{X}{Gene expression matrix with samples as rows and genes as columns.}

\item{y}{Optional vector of 0/1 indicating sample phenotype.}

\item{S}{Matrix indicating gene set membership with genes as rows and gene sets as columns. 1 indicates that the gene is a member, 0 indicates not a member.}

\item{nrot}{The number of rotations to perform. Default is 500.}

\item{fun}{Name of function for computing a gene wise test statistic.}
}
\value{
\item{p.values }{A p-value for each gene set}
\item{q.values }{A FDR q-value for each gene set}
}
\description{
GSEA with rotation test
}
\references{
Dorum, G., Snipen, L., Solheim, M. and Sabo (2009) Rotation Testing in Gene Set Enrichment
Analysis for Small Direct Comparison Experiments. \emph{Statistical Applications in Genetics
 and Molecular Biology}, \bold{8}(1), article 34.
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.
and Mesirov, J. P (2005) Gene set enrichment analysis: A knowledge-based
approach for interpreting genome-wide expression profiles, \emph{PNAS}, \bold{102},
15545-15550.
}
\author{
Guro Dorum
}
