\name{GSEArot-package}
\alias{GSEArot-package}
\alias{GSEArot}
\docType{package}
\title{
Gene Set Enrichment Analysis with rotation testing
}
\description{
This GSEA is similar to the original GSEA proposed by Subramanian et al. (2005).
The difference lies in the fact that rotation testing is replacing the permutation testing.
As shown by D{\o}rum et al. (2009) rotation testing has higher power than permutation testing
in case of small sample sizes.
}
\details{
\tabular{ll}{
Package: \tab GSEArot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Guro D{\o}rum

Maintainer: <guro.dorum@umb.no>

}
\references{
D{\o}rum, G., Snipen, L., Solheim, M. and S{\ae}b{\o} (2009). Rotation Testing in Gene Set Enrichment
Analysis for Small Direct Comparison Experiments. \emph{Statistical Applications in Genetics
and Molecular Biology}, \bold{8}(1), article 34.

Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L.,
Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S.
and Mesirov, J. P (2005) Gene set enrichment analysis: A knowledge-based
approach for interpreting genome-wide expression pro?les, \emph{PNAS}, \bold{102},
15545-15550.
}
\keyword{ package }
