% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t90th.R
\name{t90th}
\alias{t90th}
\title{Moderated gene wise t-statistics.}
\usage{
t90th(X, y = NULL)
}
\arguments{
\item{X}{Gene expression matrix with samples as rows and genes as columns}

\item{y}{Optional vector of 0/1 indicating sample phenotype}
}
\value{
Vector with one smoothed t-value for each gene
}
\description{
Computes gene wise t-statistic with variance stabilised by adding the 90th percentile.
}
\details{
Computes a gene wise t-statistic with variance stabilised by adding the 90th percentile 
of all the genes' standard deviations to a gene's standard devitation (Efron et al. 2001).
If y is given, a two-sample test is performed, otherwise a one-sample test is performed.
For a two-sample t-test the approach of Pan et al. (2003) is used.
}
\references{
Efron, B., Tibshirani, R., Storey, J. D. and Tusher, V. (2001) 
Empirical Bayes analysis of a microarray experiment, 
\emph{J Amer Statist Assoc}, 
\bold{96}, 1151-1160. 
Pan,W., Lin,J. and Le,C.T. (2003)
A mixture model approach to detecting differentially expressed genes with microarray data, 
\emph{Funct Integr Genomics}, 
\bold{3}, 117-124.
}
\author{
Guro Dorum
}
